refactor(analysis test): separate function tests
